# docs/swagger.yaml
definitions:
...
  handler.appCodeAppSecretSerializer:
    properties:
      app_code:
        type: string
      app_secret:
        type: string
    type: object
...
  handler.credentialsVerifyResponseSerializer:
    properties:
      valid:
        type: boolean
    type: object
  handler.credentialsVerifySerializer:
    properties:
      data:
        $ref: '#/definitions/handler.appCodeAppSecretSerializer'
        type: object
      type:
        type: string
    type: object
...
  handler.getMaxPolicyIDResponse:
    properties:
      id:
        type: integer
    type: object
...
  handler.listPolicyIDsResponse:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  handler.policyGetResponse:
    properties:
      action:
        $ref: '#/definitions/handler.policyResponseAction'
        type: object
      expired_at:
        example: 4102444800
        type: integer
      expression:
        additionalProperties: true
        type: object
      id:
        example: 100
        type: integer
      subject:
        $ref: '#/definitions/handler.policyResponseSubject'
        type: object
      system:
        example: bk_test
        type: string
      version:
        example: "1"
        type: string
    type: object
...
  handler.policyListResponse:
    properties:
      count:
        example: 120
        type: integer
      metadata:
        $ref: '#/definitions/handler.policyListResponseMetadata'
        type: object
      results:
        items:
          $ref: '#/definitions/handler.thinPolicyResponse'
        type: array
    type: object
  handler.policyListResponseMetadata:
    properties:
      action:
        $ref: '#/definitions/handler.policyResponseAction'
        type: object
      system:
        example: bk_test
        type: string
      timestamp:
        example: 1592899208
        type: integer
    type: object
  handler.policyResponseAction:
    properties:
      id:
        example: edit
        type: string
    type: object
  handler.policyResponseSubject:
    properties:
      id:
        example: admin
        type: string
      name:
        example: Administer
        type: string
      type:
        example: user
        type: string
    type: object
...
  handler.thinPolicyResponse:
    properties:
      expired_at:
        example: 4102444800
        type: integer
      expression:
        additionalProperties: true
        type: object
      id:
        example: 100
        type: integer
      subject:
        $ref: '#/definitions/handler.policyResponseSubject'
        type: object
      version:
        example: "1"
        type: string
    type: object
...
  util.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
info:
  contact: {}
  description: 蓝鲸权限中心后台服务 API 文档
  license: {}
  title: IAM API
  version: "1.0"
paths:
  /api/v1/engine/credentials/verify:
    post:
      consumes:
      - application/json
      description: check app code secret exists
      operationId: api-engine-check-appcode-secret-exists
      parameters:
      - description: the credential request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.credentialsVerifySerializer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: the request id
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.credentialsVerifyResponseSerializer'
              type: object
      security:
      - AppCode: []
      - AppSecret: []
      summary: Credentials Verify/认证信息验证
      tags:
      - engine
  /api/v1/engine/policies:
    get:
      consumes:
      - application/json
      description: list policies of ids or between min-max ids
      operationId: api-engine-policies-list
      parameters:
      - in: query
        name: timestamp
        type: integer
      - in: query
        name: min_id
        type: integer
      - in: query
        name: max_id
        type: integer
      - in: query
        name: ids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: the request id
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.policyListResponse'
              type: object
      security:
      - AppCode: []
      - AppSecret: []
      summary: policy list
      tags:
      - engine
  /api/v1/engine/policies/ids:
    get:
      consumes:
      - application/json
      description: list policy pks by condition
      operationId: api-engine-policy-id-list
      parameters:
      - in: query
        name: begin_updated_at
        type: integer
      - in: query
        name: end_updated_at
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: the request id
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.listPolicyIDsResponse'
              type: object
      security:
      - AppCode: []
      - AppSecret: []
      summary: policy ids list
      tags:
      - engine
  /api/v1/engine/policies/ids/max:
    get:
      consumes:
      - application/json
      description: get max policy id by condition
      operationId: api-engine-policy-id-max
      parameters:
      - in: query
        name: updated_at
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: the request id
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.getMaxPolicyIDResponse'
              type: object
      security:
      - AppCode: []
      - AppSecret: []
      summary: policy max id
      tags:
      - engine
...
  /api/v1/systems/{system_id}/policies/{policy_id}:
    get:
      consumes:
      - application/json
      description: get a policy
      operationId: api-open-system-policies-get
      parameters:
      - description: System ID
        in: path
        name: system_id
        required: true
        type: string
      - description: Policy ID
        in: path
        name: policy_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: the request id
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.policyGetResponse'
              type: object
      security:
      - AppCode: []
      - AppSecret: []
      summary: policy get
      tags:
      - open
...
securityDefinitions:
  AppCode:
    in: header
    name: X-Bk-App-Code
    type: apiKey
  AppSecret:
    in: header
    name: X-Bk-App-Secret
    type: apiKey
swagger: "2.0"