# pkg/api/engine/handler/credentials_slz.go
/*
 * TencentBlueKing is pleased to support the open source community by making 蓝鲸智云-权限中心(BlueKing-IAM) available.
 * Copyright (C) 2017-2021 THL A29 Limited, a Tencent company. All rights reserved.
 * Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at http://opensource.org/licenses/MIT
 * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
 * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations under the License.
 */

package handler

type credentialsVerifySerializer struct {
	Type string                     `json:"type"  binding:"required"`
	Data appCodeAppSecretSerializer `json:"data"  binding:"required"`
}

type appCodeAppSecretSerializer struct {
	AppCode   string `json:"app_code"  binding:"required"`
	AppSecret string `json:"app_secret"  binding:"required"`
}

type credentialsVerifyResponseSerializer struct {
	Valid bool `json:"valid"`
}